requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-qdrant==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.37
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.5.0
orjson==3.11.3
packaging==25.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pypdf==6.1.2
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
